requirements.txt
# Core Telegram Bot Framework
python-telegram-bot==20.7

# Google APIs Integration
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Task Scheduling for Memory Management
APScheduler==3.10.4

# Additional Dependencies
requests==2.31.0
python-dotenv==1.0.0